multi_agent_app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
requests==2.31.0
google-generativeai==0.4.1
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
